Fix issue where all cells are blank
There was a problem where if all cells where blank and the IgnoreEmpty flag was on, there would be nothing returned from the function. I fixed this by giving `OutStr` an initial value of "" and also checked if it had any length before returning a result. If it has no length, just pass it as is, if it has length then pass all but the last character